requirements.txt
Flask==2.3.2
twilio==8.0.0
gunicorn==20.1.0
pymongo==4.5.0
dnspython==2.3.0
Werkzeug==2.3.7
anthropic>=0.2.2
python-dotenv==1.0.0
motor==3.3.2
aiohttp==3.9.1
asgiref
uvicorn
fastapi
python-multipart
requests